requirements.txt
insightface==0.7.3
supabase==2.0.3
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
tqdm==4.66.1
pillow==10.1.0
requests==2.31.0
numpy==1.26.1
opencv-python==4.8.1.78
scipy==1.11.3
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0